Limit repo listing to find the timelines
