+++ src/backups.py
@@ -450,25 +450,26 @@ class PostgreSQLBackups(Object):
         output, _ = self._execute_command([
             "pgbackrest",
             "repo-ls",
+            "archive",
             "--recurse",
-            "--output=json",
+            "--filter",
+            "\\.history$",
         ])
 
         repository = json.loads(output).items() if output else None
-        if repository is None:
-            return dict[str, tuple[str, str]]()
-
-        return dict[str, tuple[str, str]]({
-            datetime.strftime(
-                datetime.fromtimestamp(timeline_object["time"], UTC),
-                "%Y-%m-%dT%H:%M:%SZ",
-            ): (
-                timeline.split("/")[1],
-                timeline.split("/")[-1].split(".")[0].lstrip("0"),
-            )
-            for timeline, timeline_object in repository
-            if timeline.endswith(".history") and not timeline.endswith("backup.history")
-        })
+        output = dict[str, tuple[str, str]]()
+        if repository:
+            for timeline, timeline_object in repository:
+                if not timeline.endswith("backup.history"):
+                    # 0 is the stanza -1 is the timeline file
+                    path = timeline.split("/")
+                    output[
+                        datetime.strftime(
+                            datetime.fromtimestamp(timeline_object["time"], UTC),
+                            "%Y-%m-%dT%H:%M:%SZ",
+                        )
+                    ] = (path[0], path[-1].split(".")[0].lstrip("0"))
+        return output
 
     def _get_nearest_timeline(self, timestamp: str) -> tuple[str, str] | None:
         """Finds the nearest timeline or backup prior to the specified timeline.
@@ -957,12 +958,11 @@ Stderr:
         elif is_backup_id_timeline:
             restore_stanza_timeline = timelines[backup_id]
         else:
-            backups_list = list(self._list_backups(show_failed=False).values())
-            timelines_list = self._list_timelines()
+            backups_list = list(backups.values())
             if (
                 restore_to_time == "latest"
-                and timelines_list is not None
-                and max(timelines_list.values() or [backups_list[0]]) not in backups_list
+                and timelines is not None
+                and max(timelines.values() or [backups_list[0]]) not in backups_list
             ):
                 error_message = "There is no base backup created from the latest timeline"
                 logger.error(f"Restore failed: {error_message}")
