+++ tests/integration/test_backups_aws.py
@@ -7,10 +7,10 @@ import pytest
 from pytest_operator.plugin import OpsTest
 from tenacity import Retrying, stop_after_attempt, wait_exponential
 
+from .backup_helpers import backup_operations
 from .conftest import AWS
 from .helpers import (
     DATABASE_APP_NAME,
-    backup_operations,
     db_connect,
     get_password,
     get_primary,
@@ -27,7 +27,6 @@ FAILED_TO_INITIALIZE_STANZA_ERROR_MESSAGE = "failed to initialize stanza, check 
 S3_INTEGRATOR_APP_NAME = "s3-integrator"
 tls_certificates_app_name = "self-signed-certificates"
 tls_channel = "1/stable"
-tls_config = {"ca-common-name": "Test CA"}
 
 logger = logging.getLogger(__name__)
 
@@ -43,7 +42,6 @@ async def test_backup_aws(ops_test: OpsTest, charm, aws_cloud_configs: tuple[dic
         charm,
         S3_INTEGRATOR_APP_NAME,
         tls_certificates_app_name,
-        tls_config,
         tls_channel,
         credentials,
         AWS,
