+++ tests/integration/test_backups_gcp.py
@@ -10,10 +10,10 @@ from lightkube.resources.core_v1 import Pod
 from pytest_operator.plugin import OpsTest
 from tenacity import Retrying, stop_after_attempt, wait_exponential
 
+from .backup_helpers import backup_operations
 from .conftest import GCP
 from .helpers import (
     DATABASE_APP_NAME,
-    backup_operations,
     build_and_deploy,
     cat_file_from_unit,
     db_connect,
@@ -30,7 +30,6 @@ FAILED_TO_INITIALIZE_STANZA_ERROR_MESSAGE = "failed to initialize stanza, check 
 S3_INTEGRATOR_APP_NAME = "s3-integrator"
 tls_certificates_app_name = "self-signed-certificates"
 tls_channel = "1/stable"
-tls_config = {"ca-common-name": "Test CA"}
 
 logger = logging.getLogger(__name__)
 
@@ -46,7 +45,6 @@ async def test_backup_gcp(ops_test: OpsTest, charm, gcp_cloud_configs: tuple[dic
         charm,
         S3_INTEGRATOR_APP_NAME,
         tls_certificates_app_name,
-        tls_config,
         tls_channel,
         credentials,
         GCP,
